# openapi.yaml
# The wire format of the agent HTTP interface — normative.
# Behavioral requirements (terminal-state guarantee, model interaction
# rules R1-R7, tool invocation) live in SPEC.md.
openapi: 3.1.0
info:
  title: agent-koans Agent HTTP Interface
  version: 0.1.0-draft
  description: >
    The HTTP surface every agent under test MUST expose. The harness is
    the only client. Unknown request fields MUST be ignored; response
    objects MAY carry additional fields, which the harness ignores.

paths:
  /health:
    get:
      operationId: getHealth
      summary: Readiness probe
      description: >
        MUST return 200 once the agent can accept runs. The response body
        is unspecified and ignored by the harness.
      responses:
        "200":
          description: The agent is ready.

  /runs:
    post:
      operationId: createRun
      summary: Submit a prompt
      description: >
        Accepts a prompt and starts a run asynchronously. The harness polls
        `GET /runs/{run_id}` until the run reaches a terminal state.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunRequest"
      responses:
        "201":
          description: Run created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunCreated"
        "202":
          description: Run accepted. Equivalent to 201.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunCreated"
        "400":
          description: Malformed body (not JSON, or `prompt` missing).

  /runs/{run_id}:
    get:
      operationId: getRun
      summary: Read run state
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Current run state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "404":
          description: Unknown `run_id`.

  /runs/{run_id}/abort:
    post:
      operationId: abortRun
      summary: Request cancellation of a run
      description: >
        A run still in progress MUST then reach the terminal state
        `aborted` in finite time. A run that has already reached a
        terminal state MUST keep it — a late abort never rewrites a
        committed result. Repeated aborts are idempotent (SPEC.md §3).
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Abort accepted.
        "200":
          description: Abort accepted. Equivalent to 202.
        "404":
          description: Unknown `run_id`.

  /runs/{run_id}/prompts:
    post:
      operationId: sendPrompt
      summary: Deliver a follow-up prompt to the conversation
      description: >
        Continues an existing run's conversation with a new prompt
        (SPEC.md §6.5). Delivering to a run already in a terminal state
        MUST re-open it: `status` returns to `running`, and the run MUST
        reach a terminal state again in finite time, with `output`
        carrying this turn's final answer.
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt:
                  type: string
                  minLength: 1
                  description: The follow-up prompt, verbatim.
      responses:
        "202":
          description: Prompt accepted.
        "200":
          description: Prompt accepted. Equivalent to 202.
        "404":
          description: Unknown `run_id`.

components:
  schemas:
    RunRequest:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
          minLength: 1
          description: >-
            The user's initial prompt, verbatim. Routing to the right
            conversation (SPEC.md §6.4) matches by substring, so an empty
            prompt would match every request.
        tools:
          type: array
          default: []
          items:
            $ref: "#/components/schemas/ToolDef"
        subagents:
          type: array
          default: []
          description: >-
            Delegates available for this run (SPEC.md §6.4). The agent
            recognizes its own delegation tool call and, instead of
            invoking the tool server, starts a subagent conversation with
            the named delegate.
          items:
            $ref: "#/components/schemas/SubagentDef"
        limits:
          type: object
          description: Optional per-run budgets the agent must honor (SPEC.md §4 R5).
          properties:
            max_model_requests:
              type: integer
              minimum: 1
              description: >-
                Maximum number of model requests the agent may issue for
                this run. Exhausting the budget before a final answer must
                end the run as `aborted`.

    SubagentDef:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          description: How the model addresses this delegate in a delegation instruction.
        description:
          type: string

    ToolDef:
      type: object
      required: [name, input_schema]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        input_schema:
          type: object
          description: JSON Schema for the tool's arguments.

    RunCreated:
      type: object
      required: [run_id]
      properties:
        run_id:
          type: string
          minLength: 1

    Run:
      type: object
      required: [run_id, status]
      properties:
        run_id:
          type: string
        status:
          type: string
          enum: [running, completed, failed, aborted]
          description: >
            `running` is the only non-terminal state. Every run MUST reach
            a terminal state in finite time (SPEC.md §3.3).
        output:
          type: string
          description: Final answer. REQUIRED when `status` is `completed`.
        error:
          type: string
          description: Failure detail. SHOULD be present when `status` is `failed`.